validate urls before any web ops
