+++ autogpt/commands/git_operations.py
@@ -3,6 +3,8 @@ from git.repo import Repo
 
 from autogpt.commands.command import command
 from autogpt.config import Config
+from autogpt.url_utils.validators import validate_url
+
 
 CFG = Config()
 
@@ -14,6 +16,7 @@ CFG = Config()
     CFG.github_username and CFG.github_api_key,
     "Configure github_username and github_api_key.",
 )
+@validate_url
 def clone_repository(repository_url: str, clone_path: str) -> str:
     """Clone a GitHub repository locally.
