+++ autogpt/commands/web_playwright.py
@@ -10,6 +10,7 @@ except ImportError:
 from bs4 import BeautifulSoup
 
 from autogpt.processing.html import extract_hyperlinks, format_hyperlinks
+from autogpt.url_utils.validators import validate_url
 
 
 def scrape_text(url: str) -> str:
